% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_friends.R
\name{animal_friends}
\alias{animal_friends}
\title{Print Statement About Friends' Cats and Dogs}
\usage{
animal_friends(my_name, cat_friend, dog_friend, my_cat, cat_1, cat_2, dog_1)
}
\arguments{
\item{my_name}{A character stating your name.}

\item{cat_friend}{A character stating your cat owning friend.}

\item{dog_friend}{A character stating your dog owning friend.}

\item{my_cat}{A character stating your cat's name.}

\item{cat_1}{A character stating your friend's first cat's name.}

\item{cat_2}{A character stating your friend's second cat's name.}

\item{dog_1}{A character stating your friend's dog's name}
}
\value{

}
\description{
Print Statement About Friends' Cats and Dogs
}
