% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_friends.R
\name{cat_friends}
\alias{cat_friends}
\title{Print Statement About Friends' Cats}
\usage{
cat_friends(cat_1_0, cat_1_2, cat_2, cat_3, friend_1, friend_2, friend_3)
}
\arguments{
\item{cat_1_0}{A character stating the name of friend 1's first cat}

\item{cat_1_2}{A character stating the name of friend 1's second cat}

\item{cat_2}{A character sating the name of friend 2's cat}

\item{cat_3}{A character sating the name of friend 3's cat}

\item{friend_1}{A character stating the name of friend 1}

\item{friend_2}{A character stating the name of friend 2}

\item{friend_3}{A character stating the name of friend 3}
}
\value{

}
\description{
Print Statement About Friends' Cats
}
