% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_me.R
\name{pet_trick_checkin}
\alias{pet_trick_checkin}
\title{Checks in with your friend on pet training}
\usage{
pet_trick_checkin(name, pet_name, trick_name, practice_number)
}
\arguments{
\item{name}{A character containing your friend's name}

\item{pet_name}{A character containing your friend's pet's name}

\item{trick_name}{A character containing the name of the trick the pet is learning}

\item{practice_number}{A number of times the trick needs to be practiced}
}
\value{

}
\description{
Contains one functions, pet_trick_checkin, that prints a statement asking your friend how their pet is doing learning a new trick.
}
